chore: rename common components
