+++ src/components/common/ModalRenderer.tsx
@@ -1,9 +1,9 @@
 import modalState from "@atoms/common/modalState";
 import { useRecoilValue } from "recoil";
 
-const Modal = () => {
+const ModalRenderer = () => {
   const { Component, props } = useRecoilValue(modalState);
   return Component ? <Component {...props} /> : null;
 };
 
-export default Modal;
+export default ModalRenderer;
