+++ src/components/mypage/UserInfoSection.tsx
@@ -9,7 +9,7 @@ import {
   UserYearSpan,
 } from "./style/UserInfoSection";
 import Image from "next/image";
-import { AuthSelfHiddenMask } from "@components/common/AuthSelfMask/AuthSelfMask";
+import { AuthSelfHiddenMask } from "@components/common/AuthSelfMask";
 import mypageImg from "@/image/mypageImg";
 
 const UserInfoSection = () => {
