+++ src/pages/mypage/[id].tsx
@@ -1,4 +1,4 @@
-import AuthSelfMask from "@components/common/AuthSelfMask/AuthSelfMask";
+import AuthSelfMask from "@components/common/AuthSelfMask";
 import MypageBanner from "@components/mypage/MypageBanner";
 import MypageContainer from "@containers/mypage/MyPageContainer";
